fix bug with gcd shellcode

+++ src/test_shellcodes/gcd.c
@@ -2,7 +2,7 @@
 
 size_t _start(size_t a, size_t b) {
     while (b != 0) {
-        int temp = b;
+        size_t temp = b;
         b = a % b;
         a = temp;
     }
